Allow copying of objects that are not owned by the bucket owner.
Not all objects in an S3 bucket are necessarily owned by the bucket
owner. This occurs when getting detailed billing data from AWS dumped
into one of your buckets - whilst you have access to read the data (and
get charged for the data), you don't actually own the object which
prevents you from querying it for ACL information.

There doesn't seem to be a way in which you can query the object for
it's ownership information, so the "best" logic I could come up with is
creating a blank ACL if we are unable to get ACL perms after the max
attempts.